Added Navagation Between Financial components
